Extract AppPackage into its own, er, package
Make the AppPackage parsing and loading logic available to other
packages outside of the CLI debugger. And tweak the API a bit.

Strip out the package generation logic as it is no longer used anywhere.

Change the loader function to take a JSZip instance as an argument. This
decouples it from the actual loading of the zip file and eliminates the
direct dependency on JSZip.

Change the parsed `AppPackage` instances to be plain old data objects.
Remove the constructor.

Export the loader function individually. Future loaders and utility
functions will also be exported individually so that this package can be
tree-shaken.

Drop the deprecated `manifest` property from the parsed AppPackage
object. Any consumer that relied on that property can replicate the
functionality by reading and parsing `manifest.json` directly from the
app package ZIP file.
```js
const manifest = await packageZip.file('manifest.json')
  .async('text')
  .then(JSON.parse);
```