Break up FDBTypes.ts into more manageable modules
FDBTypes.ts has grown to nearly 1000 lines in length, making it unwieldy
to read or edit. Break it up into more reasonably-sized chunks by moving
the types for each aspect of the protocol into its own module.

There should not be any impact on the public interface of the
fdb-protocol package. All the same types are exported under the same
names from the main module of the package. The only outwardly-visible
change is the addition of some new exports for AppHost capabilities.