Test with Node 8.0 and 10.0 in CI (#30)
* Test with Node 8.0 and 10.0 in CI

The sdk-cli unit tests are broken in Node 10.14.2, presumably because of
a change that broke mock-fs. Lock the Node versions we run tests against
to 8.0.0 and 10.0.0 to both test that everything runs against the lowest
Node version we support and that these sorts of breakages don't happen
again.

Add an `engines` section to the sdk-cli package.json to be explicit
about which Node versions we support. This matches the versions we
support for fitbit-sdk-toolchain.

* Bump locked package versions so it can install on node 8.0

* Make Jest config compatible with Node 8.0

* Enable yarn cache