Update to AppAuth-JS v1
This lets us drop some direct and indirect dependencies, which is nice.

Whatever AppAuth-JS uses for making HTTP requests seems to
unconditionally add the port to the request URL, which was causing nock
to reject those requests in the unit tests since it didn't exactly match
the configured base URL. Adding the port to the auth config URLs was the
simplest fix.