memory-profiler: Convert to V8 Heap Dump
This replaces the old output format in favor of a V8 Heap Snapshot
which is usable in Chrome Dev Tools.

This will hopefully be useful in debugging heap usage even though
it is not quite as full of metadata compared to the heap snapshots taken
from a normal V8 app.

This also adds the ability to export the heap snapshot to GEXF for use
in graphing tools like Gephi.

Also included is the ability to launch a REPL to manually explore the heap
graph.

Signed-off-by: Ben Keith <keitwb@gmail.com>